tests
test

+++ src/main/java/com/evatool/requirements/events/listener/RequirementEventListener.java
@@ -6,6 +6,11 @@ import com.evatool.global.event.dimension.DimensionUpdatedEvent;
 import com.evatool.global.event.impact.ImpactCreatedEvent;
 import com.evatool.global.event.impact.ImpactDeletedEvent;
 import com.evatool.global.event.impact.ImpactUpdatedEvent;
+import com.evatool.global.event.stakeholder.StakeholderCreatedEvent;
+import com.evatool.global.event.stakeholder.StakeholderDeletedEvent;
+import com.evatool.impact.application.json.mapper.ImpactStakeholderJsonMapper;
+import com.evatool.impact.common.exception.EventEntityAlreadyExistsException;
+import com.evatool.impact.common.exception.EventEntityDoesNotExistException;
 import com.evatool.requirements.entity.RequirementDimension;
 import com.evatool.requirements.entity.RequirementsImpact;
 import com.evatool.requirements.repository.RequirementDimensionRepository;
@@ -14,12 +19,14 @@ import com.evatool.requirements.repository.RequirementsVariantsRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationEvent;
+import org.springframework.context.ApplicationListener;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 
 @Component
-public class RequirementEventListener {
+public class RequirementEventListener implements ApplicationListener<ApplicationEvent> {
 
     final Logger logger = LoggerFactory.getLogger(RequirementEventListener.class);
 
@@ -37,6 +44,9 @@ public class RequirementEventListener {
         logger.info("impact created event ");
         logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
 
+        if (requirementsImpactsRepository.existsById(RequirementsImpact.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
         requirementsImpactsRepository.save(RequirementsImpact.fromJson(event.getJsonPayload()));
     }
 
@@ -45,6 +55,9 @@ public class RequirementEventListener {
     public void impactUpdated(ImpactUpdatedEvent event) {
         logger.info("Impact updated event");
         logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        if (!requirementsImpactsRepository.existsById(RequirementsImpact.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementsImpactsRepository.save(RequirementsImpact.fromJson(event.getJsonPayload()));
     }
 
@@ -53,6 +66,9 @@ public class RequirementEventListener {
     public void impactDeleted(ImpactDeletedEvent event) {
         logger.info("Impact deleted event");
         logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        if (!requirementsImpactsRepository.existsById(RequirementsImpact.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementsImpactsRepository.delete(RequirementsImpact.fromJson(event.getJsonPayload()));
     }
     @EventListener
@@ -60,6 +76,9 @@ public class RequirementEventListener {
     public void dimensionCreated(DimensionCreatedEvent event) {
         logger.info("dimension created event");
         logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        if (requirementDimensionRepository.existsById(RequirementDimension.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
         requirementDimensionRepository.save(RequirementDimension.fromJson(event.getJsonPayload()));
     }
 
@@ -68,6 +87,9 @@ public class RequirementEventListener {
     public void dimensionUpdated(DimensionUpdatedEvent event) {
         logger.info("dimension updated event");
         logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        if (!requirementDimensionRepository.existsById(RequirementDimension.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementDimensionRepository.save(RequirementDimension.fromJson(event.getJsonPayload()));
     }
 
@@ -76,7 +98,26 @@ public class RequirementEventListener {
     public void dimensionDeleted(DimensionDeletedEvent event) {
         logger.info("dimension deleted event");
         logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        if (!requirementDimensionRepository.existsById(RequirementDimension.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementDimensionRepository.delete(RequirementDimension.fromJson(event.getJsonPayload()));
+    }
+
+    @Override
+    public void onApplicationEvent(ApplicationEvent event) {
+
+/*
+        logger.info("Event received");
+        var jsonPayload = event.getJsonPayload();
+        var stakeholder = ImpactStakeholderJsonMapper.fromJson(jsonPayload);
+        if (stakeholderRepository.existsById(stakeholder.getId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
+        stakeholderRepository.save(stakeholder);
+        logger.info("Event successfully processed");
+        */
+
     }
     //events do not exist at the moment
